Alter the repo to host multiple github actions
If we need to have a repo per action, things are
going to get messy.

Instead, have a single repo with a directory per
action.

The downside of this is that you have to use
`@master` instead of `@release` (well, you might
not have to, but it means we don't have to worry
about releases for each individual action/dir)