Don't try to parse non-existent YAML files
If a commit removes files, they will still be listed as "files
affected by this PR." Then, when we try to parse the non-existent
file, we get an error and the action shows a failure status.

This change causes the action to ignore any listed files which do
not currently exist in the repository.