Fix a syntax error in reject-malformed-yaml.rb
I have no idea why this syntax error started to appear:

    $ ruby malformed-yaml/reject-malformed-yaml.rb
    malformed-yaml/reject-malformed-yaml.rb:13: syntax error, unexpected
    rescue, expecting '}'
      rescue Psych::SyntaxError
    malformed-yaml/reject-malformed-yaml.rb:15: syntax error, unexpected
    '}', expecting end
      }.compact

I thought `{...}` and `do ... end` were equivalent in this case,
but apparently not.

It seems to have something to do with the specific version of ruby,
because `standardrb --fix` wants to revert the change if I simply
swap `{...}` and `do ... end`.

This change should be a bit more robust than that. standardrb seems
happy with it. It also makes the code easier to understand.