Bugfix: terraform12 validate was failing
As written, the code was running `terraform12 init` on the root
directory of the repository, but `terraform12 validate` on the
namespace sub-directory, in which `init` had not been run. So the
validate step was failing with "module not installed" errors.

This change should fix this, by ensuring that `init` and `validate`
are both targeting the namespace sub-directory.