Remove `terraform12 validate` step
In terraform 0.12, the validate step has no option *not* to
check variables. This means `terraform12 init` must be run before
we can run `validate`.

Doing this requires passing a lot of configuration to `terraform12
init`, which would make this action too repository-specific to be part
of a shared github action.

This commit removes the `terraform12 validate` action, if terraform
0.12 source code is detected.