Replace symbolic links with hard links
The symbolic links allowed files to be defined in
`shared/` and referenced in several different 
GitHub actions. However, following symlinks in the
`docker build` stage is unreliable (in fact, it's
not supposed to work at all).

This change replaces the symlinks with hard links.
I'm not sure if the hard links will survive a
`git push` ... `git pull`. It's possible the
copies may go out of sync. But, at least it's 
possible to build the docker images, this way.