Create a smaller docker image for code-formatter
The code-formatter github action was using the cloud platform tools
image. The advantage of this is that it guarantees that the terraform
version in this action is always the same as the "official" version used
on the Cloud Platform.

The downside is that the cloud platform tools image is (currently)
around 1.4G in size, which is a lot larger than it needs to be. This
slows down the code formatter action, and will eventually cost us money
if/when we start paying for github action execution time.

This change results in a much smaller docker image (approx. 128MB) which
should be considerably faster. However, it does mean we may have to update
the terraform version here whenever it changes on the cloud platform.

fixes https://github.com/ministryofjustice/operations-engineering/issues/44