Create a getOrigin function to decide which cluster to inspect
We have a situation as follows:
- two cluster directories in the environments repository; one live and
one live-1
- the namespace might live in one of the clusters, both or none.

if it lives on both or either, the `getOrigin` function will first
identify if it needs to check `live` or `live-1`. If a response code
returns anything other than a 200, the function assumes it's a new
namespace. In which case it'll create a new set of options and check the
PR branch.

This is not an ideal situation and there is some code that doesn't
follow the DRY principle.