Loop over all teams in a namespace and check if user is a member
This is the crux of the application, we need to know if the github user
(who raised the PR) is a member any of the teams outlined in a namespace
rbac file.

This loop checks each namespace and calls isUserValid which will return
a bool if the user is or isn't in the required rbac groups.