Amend isUserValid function to return a bool if user can merge
Instead of looping over every team in the `ministryofjustice`
orginisation, it seemed considerably more responsible to list all team
members using the namespace as a slug. This means we can loop over all
teams in a namespace, make a call to the GH api for team members and
check if our user is a member.

This uses nested loops so it's complexity is O(n^2), so not ideal. I
could move this to graphql and only make one GH rest API call, but for
now I'll leave it as it is.