Add Terraform static code analysis action
This action was created to solve the Modernisation Platform problem of
how to run Terraform static analysis checks in repositories with
multiple Terraform folders.

Most Terraform static analysis and linters are design to run the same
way as Terraform - in one folder.  This means they generally do not have
the ability to scan multiple folders or recursively.

On the Modernisation Platform we have multiple Terraform folders and
nested folders, to add existing actions to do static analysis we would
have to add a workflow for each Terraform folder which would quickly
become messy, a lot of code to manage and hard to keep in sync. Not to
mention that we automatically generate new Terraform for each new
account.

The solution here is to create an action which iterates through each
terraform folder you want to check and runs each test in the folder. You
can choose to run the action on a single folder, on all folders, or on
folders which have had changes to them.  Eg we have the workflow set up
to run on changed folders when a PR is raised, but do a full scan of all
folders when triggered via the workflow dispatch:
https://github.com/ministryofjustice/modernisation-platform/blob/main/.github/workflows/terraform-static-analysis.yml

This gives us the ability to scan our entire repo, but have a more
focused scan when doing individual sections of work.

This is initial starting code and can definitely be cleaned up and
improved, a few known things than can be improved at a later date if
needed:

 - Clean up docker file, currently 3 different tools are being installed
an this could probably be done in a better way to reduce the image build
time.
 - Add a flag to turn of Terraform init if not needed
 - Add ability to choose other versions of tools (TFsec has this but not
the others)