Remove terraform init and the need for credentials
Since we are using 3rd party tools, there is a chance that malicious
scripts could be injected into them.  Removing the Terraform init as
this is not really needed, (some checks will not be done on terraform
modules but this can be done on the module directly if needed) without
terraform init there is no need to pass in the AWS credentials.

Also redirecting stderr to stdout so that you see those messages along
with the relevant tests being run (running in a subshell and waiting was
causing these errors to show out of context at the top of the run).