Add separate config for tflint AWS plugin
The AWS plugin is currently not working if a local doesn't have a valid
value if it's generate dynamically, there is no way to ignore this
error, so separating the tflint config so the default is no plugins, and
then the AWS plugin is defined in the tflint.aws.hcl file.

This makes more sense anyway, and when the aws plugin is working again
we can change actions to use that.  I've checked the usage of this and
currently only the modernisation platform is using the default, DSO uses
the Azure one.