:arrow_up: Pin checkout action to a specific commit
This commit pins the 'checkout' GitHub action across multiple workflows to a specific commit hash (b4ffde65f46336ab88eb53be808477a3936bae11). The change provides more control over the versioning and reliability of the codebase by ensuring a specific, stable version of the action is used. This could be particularly useful if there are breaking changes or bugs in later versions that we want to avoid, ensuring our workflow remains consistent and predictable.