:sparkles: Add Github action to reject malformed YAML files
This commit introduces a new GitHub Action that runs on every pull request opening or synchronization. It utilizes the `malformed-yaml` action from the Ministry of Justice's collection of GitHub Actions, which flags any badly structured YAML files in the PR.

The action is configured to work with the necessary GitHub environment variables for access and identification: GITHUB_TOKEN, PR_NUMBER, and REPOSITORY_NAME. This ensures smooth operation in context of different pull requests and repositories.